comments added and. mongo relationship done using DBRef

+++ src/main/java/com/aspire/usermanage/usermanagement/repository/GroupRepository.java
@@ -1,13 +1,19 @@
 package com.aspire.usermanage.usermanagement.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import com.aspire.usermanage.usermanagement.modal.Group;
 
+/**
+ * Group Repsoitory on mongo template
+ * 
+ * @author kathiravan.sethurama
+ *
+ */
 @Repository
-public interface GroupRepository extends MongoRepository<Group, Long>{
+public interface GroupRepository extends MongoRepository<Group, Integer>{
 
 }
