+++ src/main/java/com/aspire/usermanage/usermanagement/repository/UserRepository.java
@@ -1,13 +1,18 @@
 package com.aspire.usermanage.usermanagement.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import com.aspire.usermanage.usermanagement.modal.User;
-
+/**
+ * User Repsoitory on mongo template
+ * 
+ * @author kathiravan.sethurama
+ *
+ */
 @Repository
-public interface UserRepository extends MongoRepository<User,Long>{
+public interface UserRepository extends MongoRepository<User,Integer>{
